alpaca-to-influxdb/requirements.txt
websockets
influxdb-client
pyyaml